Add support for new database variants
- Add `db::Type` enum: holds supported database variants
- Add `db::Databse` enum: holds database connection
- Add `db::Inserter` trait: defines behavior of database inserter
